requirements.txt
bcrypt==4.2.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
filelock==3.16.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.7.1
fsspec==2024.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.2.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.11.6
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
setuptools==75.6.0
six==1.16.0
smart-open==7.0.5
sympy==1.13.1
threadpoolctl==3.5.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.3
wrapt==1.16.0
gunicorn
